style: [DDO-9] bem notification and notice, fixed styling notifications
